fix: background job utils - move background job utils to separate file
